♻️ FA:P 아카이브와 패션 전체보기 뷰 분리
